EAMxx: remove unnecessary atmosphere_radiation_yakl.*pp
